requirements.txt
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.10.0
blinker==1.4
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.1
flasgger==0.9.7.1
Flask==2.1.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
gcloud==0.18.3
googleapis-common-protos==1.63.0
greenlet==3.0.3
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
MarkupSafe==2.1.5
meson==1.3.1
mistune==3.0.2
more-itertools==8.10.0
mysqlclient==2.2.4
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
packaging==24.0
pluggy==1.5.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.8.0
pycryptodome==3.20.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
PyMySQL==1.1.0
pyparsing==2.4.7
Pyrebase4==4.8.0
pytest==8.2.1
python-apt==2.4.0+ubuntu3
python-dotenv==1.0.1
python-jwt==4.1.0
PyYAML==5.4.1
referencing==0.35.1
requests==2.31.0
requests-toolbelt==0.10.1
rpds-py==0.18.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
spark-parser==1.8.9
SQLAlchemy==2.0.30
systemd-python==234
tomli==2.0.1
typing_extensions==4.11.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uncompyle6==3.9.0
urllib3==1.26.18
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.1.1
xdis==6.0.5
zipp==1.0.0
